removed utilities file and moved to lib folder inside utilis.js
